+++ build-packages.sh
@@ -40,11 +40,4 @@ cat pkg/small.robin \
     pkg/list.robin \
     pkg/env.robin \
     pkg/misc.robin \
-    > pkg/stdlib-no-builtins.robin
-
-cat pkg/boolean.robin \
-    pkg/arith.robin \
-    pkg/list.robin \
-    pkg/env.robin \
-    pkg/misc.robin \
     > pkg/stdlib.robin